etl-lambda/my_function/requirements.txt
pandas
boto3
sqlalchemy
python-dotenv
psycopg2-binary
